msm8998-common: doze: Treat ProximitySensor values as int
The maximum range is reported as 5.000305 but the distance is just 5.0
which leads to isNear being always true. Cast them to int because the
only reported values are 0 and 5 anyway to get the isNear variable correct.

+++ doze/src/org/lineageos/settings/doze/ProximitySensor.java
@@ -60,7 +60,7 @@ public class ProximitySensor implements SensorEventListener {
 
     @Override
     public void onSensorChanged(SensorEvent event) {
-        boolean isNear = event.values[0] < mSensor.getMaximumRange();
+        boolean isNear = (int) event.values[0] < (int) mSensor.getMaximumRange();
         if (mSawNear && !isNear) {
             if (shouldPulse(event.timestamp)) {
                 DozeUtils.wakeOrLaunchDozePulse(mContext);
